* Add Backbone models to front-end.
* Change contest creation to use main web-app Games model.

* Create contests and games for multiple days by grabbing game schedules during
  overnight stats run.

* Populate gamecenter with current game state during initial page-load,
  so that even things that aren't changing are seen.

Currently, real-time updates are broken because of the change in the front-end
models. A page-reload is required to see current state. However, I'd like to get
it in because there is a lot of logic in the update. From here it shouldn't be
too hard to wire up the backbone models to incoming events from the pusher.