Record contest outcomes
When games finish, check entries and record their completion if not completed.
When all entries in a contest are done, assign final positions to them.
This way, entries in a completed contest will have their positions and final-scores
annotated as the canonical 'result' of the contest.

Do this calculation in the realtime worker loop, so we update entries and contests
as they complete.