Fix simulated-game data so there are no players in multiple games.
The simulated-game data is actually a combination of data over multiple days.
Some players were therefore in 2 games on the same day in simulation.

At the moment our web app isn't robust to a player in 2 games on the same day.
So for now, remove the offending games from simulation, so that any player
is in only one game. This removed 4 out of 16 games from the sim.

The 'seed' logic also removes the offending games from the DB if they were
previously present. This way a simple 'rake db:seed' on any DB should correct
its state.